Did it wher TContext : MonoBehavior
Think: what are the cases where t might not be monobehavior?

Because honestly I wish I should just pass around context easily...

the problem is that I could write contextless states but then they couldn't reference it in their own code. But maybe I can encapsulate it as a namespace?